Remove kubeproxy dependencies from kubernetes repo
