+++ pkg/api/v1/service/util.go
@@ -66,41 +66,3 @@ func GetLoadBalancerSourceRanges(service *v1.Service) (utilnet.IPNetSet, error) 
 	}
 	return ipnets, nil
 }
-
-// RequestsOnlyLocalTraffic checks if service requests OnlyLocal traffic.
-func RequestsOnlyLocalTraffic(service *v1.Service) bool {
-	if service.Spec.Type != v1.ServiceTypeLoadBalancer &&
-		service.Spec.Type != v1.ServiceTypeNodePort {
-		return false
-	}
-	return service.Spec.ExternalTrafficPolicy == v1.ServiceExternalTrafficPolicyTypeLocal
-}
-
-// RequestsOnlyLocalTrafficForInternal checks if service prefers Node Local
-// endpoints for internal traffic
-func RequestsOnlyLocalTrafficForInternal(service *v1.Service) bool {
-	if service.Spec.InternalTrafficPolicy == nil {
-		return false
-	}
-	return *service.Spec.InternalTrafficPolicy == v1.ServiceInternalTrafficPolicyLocal
-}
-
-// NeedsHealthCheck checks if service needs health check.
-func NeedsHealthCheck(service *v1.Service) bool {
-	if service.Spec.Type != v1.ServiceTypeLoadBalancer {
-		return false
-	}
-	return RequestsOnlyLocalTraffic(service)
-}
-
-// GetServiceHealthCheckPathPort returns the path and nodePort programmed into the Cloud LB Health Check
-func GetServiceHealthCheckPathPort(service *v1.Service) (string, int32) {
-	if !NeedsHealthCheck(service) {
-		return "", 0
-	}
-	port := service.Spec.HealthCheckNodePort
-	if port == 0 {
-		return "", 0
-	}
-	return "/healthz", port
-}
