+++ pkg/api/v1/service/util_test.go
@@ -128,116 +128,3 @@ func TestAllowAll(t *testing.T) {
 	checkAllowAll(true, "192.168.0.0/0")
 	checkAllowAll(true, "192.168.0.1/32", "0.0.0.0/0")
 }
-
-func TestRequestsOnlyLocalTraffic(t *testing.T) {
-	checkRequestsOnlyLocalTraffic := func(requestsOnlyLocalTraffic bool, service *v1.Service) {
-		res := RequestsOnlyLocalTraffic(service)
-		if res != requestsOnlyLocalTraffic {
-			t.Errorf("Expected requests OnlyLocal traffic = %v, got %v",
-				requestsOnlyLocalTraffic, res)
-		}
-	}
-
-	checkRequestsOnlyLocalTraffic(false, &v1.Service{})
-	checkRequestsOnlyLocalTraffic(false, &v1.Service{
-		Spec: v1.ServiceSpec{
-			Type: v1.ServiceTypeClusterIP,
-		},
-	})
-	checkRequestsOnlyLocalTraffic(false, &v1.Service{
-		Spec: v1.ServiceSpec{
-			Type: v1.ServiceTypeNodePort,
-		},
-	})
-	checkRequestsOnlyLocalTraffic(false, &v1.Service{
-		Spec: v1.ServiceSpec{
-			Type:                  v1.ServiceTypeNodePort,
-			ExternalTrafficPolicy: v1.ServiceExternalTrafficPolicyTypeCluster,
-		},
-	})
-	checkRequestsOnlyLocalTraffic(true, &v1.Service{
-		Spec: v1.ServiceSpec{
-			Type:                  v1.ServiceTypeNodePort,
-			ExternalTrafficPolicy: v1.ServiceExternalTrafficPolicyTypeLocal,
-		},
-	})
-	checkRequestsOnlyLocalTraffic(false, &v1.Service{
-		Spec: v1.ServiceSpec{
-			Type:                  v1.ServiceTypeLoadBalancer,
-			ExternalTrafficPolicy: v1.ServiceExternalTrafficPolicyTypeCluster,
-		},
-	})
-	checkRequestsOnlyLocalTraffic(true, &v1.Service{
-		Spec: v1.ServiceSpec{
-			Type:                  v1.ServiceTypeLoadBalancer,
-			ExternalTrafficPolicy: v1.ServiceExternalTrafficPolicyTypeLocal,
-		},
-	})
-}
-
-func TestNeedsHealthCheck(t *testing.T) {
-	checkNeedsHealthCheck := func(needsHealthCheck bool, service *v1.Service) {
-		res := NeedsHealthCheck(service)
-		if res != needsHealthCheck {
-			t.Errorf("Expected needs health check = %v, got %v",
-				needsHealthCheck, res)
-		}
-	}
-
-	checkNeedsHealthCheck(false, &v1.Service{
-		Spec: v1.ServiceSpec{
-			Type: v1.ServiceTypeClusterIP,
-		},
-	})
-	checkNeedsHealthCheck(false, &v1.Service{
-		Spec: v1.ServiceSpec{
-			Type:                  v1.ServiceTypeNodePort,
-			ExternalTrafficPolicy: v1.ServiceExternalTrafficPolicyTypeCluster,
-		},
-	})
-	checkNeedsHealthCheck(false, &v1.Service{
-		Spec: v1.ServiceSpec{
-			Type:                  v1.ServiceTypeNodePort,
-			ExternalTrafficPolicy: v1.ServiceExternalTrafficPolicyTypeLocal,
-		},
-	})
-	checkNeedsHealthCheck(false, &v1.Service{
-		Spec: v1.ServiceSpec{
-			Type:                  v1.ServiceTypeLoadBalancer,
-			ExternalTrafficPolicy: v1.ServiceExternalTrafficPolicyTypeCluster,
-		},
-	})
-	checkNeedsHealthCheck(true, &v1.Service{
-		Spec: v1.ServiceSpec{
-			Type:                  v1.ServiceTypeLoadBalancer,
-			ExternalTrafficPolicy: v1.ServiceExternalTrafficPolicyTypeLocal,
-		},
-	})
-}
-
-func TestRequestsOnlyLocalTrafficForInternal(t *testing.T) {
-	checkRequestsOnlyLocalTrafficForInternal := func(expected bool, service *v1.Service) {
-		res := RequestsOnlyLocalTrafficForInternal(service)
-		if res != expected {
-			t.Errorf("Expected internal local traffic = %v, got %v",
-				expected, res)
-		}
-	}
-
-	// default InternalTrafficPolicy is nil
-	checkRequestsOnlyLocalTrafficForInternal(false, &v1.Service{})
-
-	local := v1.ServiceInternalTrafficPolicyLocal
-	checkRequestsOnlyLocalTrafficForInternal(true, &v1.Service{
-		Spec: v1.ServiceSpec{
-			InternalTrafficPolicy: &local,
-		},
-	})
-
-	cluster := v1.ServiceInternalTrafficPolicyCluster
-	checkRequestsOnlyLocalTrafficForInternal(false, &v1.Service{
-		Spec: v1.ServiceSpec{
-			InternalTrafficPolicy: &cluster,
-		},
-	})
-}
