+++ pkg/proxy/util/utils.go
@@ -30,7 +30,7 @@ import (
 	utilrand "k8s.io/apimachinery/pkg/util/rand"
 	"k8s.io/apimachinery/pkg/util/sets"
 	"k8s.io/client-go/tools/record"
-	helper "k8s.io/kubernetes/pkg/apis/core/v1/helper"
+	helper "k8s.io/component-helpers/networking/corev1"
 	utilsysctl "k8s.io/kubernetes/pkg/util/sysctl"
 	utilnet "k8s.io/utils/net"
