+++ staging/src/k8s.io/component-helpers/networking/OWNERS
@@ -1,0 +1,8 @@
+# See the OWNERS docs at https://go.k8s.io/owners
+
+approvers:
+- sig-network-approvers
+reviewers:
+- sig-network-reviewers
+labels:
+- sig/network
